Added alignedCoordinateChange
It is now possible to change the frame of a vector space using
Point.alignedCoordinateChange.

It does not support rotation of the frame, though. Giving it "bad"
values (i.e. two initial points with the same x or y, or just two
identical points) will result in an IllegalArgumentException since no
information can be gathered from such data.